nfl_sim Phase 5C-1: shared solver (D46), CRPS fix (D47), td_player_id labels (D48), grader void rule (D49), engine QB identity (D50), SGP raking, sit-key fix
Code changes only — docs in next commit.

1. Shared anchoring solver: run_anchored_chunked in anchor.py with params
   from anchor block in params_v1.json. run_week.py imports it.
2. CRPS: removed spurious /n in pairwise term.
3. actuals.py: single source of truth for player stats, using td_player_id.
   calibration.py and grade_week.py both delegate to it.
4. Grader: player with no stat rows but active grades actual=0, not void.
5. Engine QB identity: uses is_starting_qb flag (raises for 2026+ if missing,
   falls back to depth_order for historical seasons).
6. SGP raking: sgp_probability_raked with iterative proportional fitting.
7. Situational tendency key: int(down).astype(str) in builder (was float str).
8. params_v1.json: anchor block with n_sims, chunk_size, max_iter,
   damp_limit_pts, J_INV, J_FWD.

Tests: test_5c1.py 8/8 pass.

+++ nfl/sim/ratings.py
@@ -501,7 +501,7 @@ def build_situational_proe(scrimmage_plays, params, league_means):
     pass_oe is in nflverse units: percentage points above league expected pass rate."""
     k_t = params.get("k_tendency", 200)
     scrim = scrimmage_plays.copy()
-    scrim["down_b"] = scrim["down"].astype(str)
+    scrim["down_b"] = scrim["down"].astype(int).astype(str)
     scrim["dist_b"] = pd.cut(scrim["ydstogo"], bins=[0, 3, 7, 100],
                               labels=["short", "med", "long"], right=True)
     scrim["score_b"] = pd.cut(scrim["score_differential"], bins=[-100, -9, 8, 100],
